Rename GUI to CLI
Since we are adding a new GUI, we can now refer to this as the CLI (Command Line Interface)
